+++ source/extensions/lightspeed.trex.layout.stagecraft/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.4.2]
+### Changed
+- Unsubscribe from the Feature Flag core on destroy
+
 ## [1.4.1]
 ### Changed
 - Moved the Stage Manager behind a feature flag