+++ source/extensions/omni.flux.feature_flags.core/docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0]
+### Added
+- Added unit tests
+- Added `unsubscribe_feature_flags_changed` to cleanup settings subscriptions
+- Added `feature_flags_changed` context manager to manage settings subscriptions
+
 ## [1.0.0]
 ### Added
 - Init commit.