+++ source/extensions/omni.flux.feature_flags.widget/docs/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.1]
+### Added
+- Added identifiers to UI elements
+- Added unit tests for the tree Model & Items
+
+### Changed
+- Unsubscribe from the Feature Flag core on destroy
+
 ## [1.0.0]
 ### Added
 - Init commit.